+++ css-packs/css-reset/examples/constellation/css-reset-simple-tables.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+const CSSResetSimpleTablesExample = () => {
+  return (
+    <div>
+      <table>
+        <thead>
+          <tr>
+            <th>Item</th>
+            <th>Qty</th>
+            <th>Price</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td>Apple</td>
+            <td>3</td>
+            <td>$5.42</td>
+          </tr>
+          <tr>
+            <td>Orange</td>
+            <td>6</td>
+            <td>$4.60</td>
+          </tr>
+          <tr>
+            <td>Banana</td>
+            <td>12</td>
+            <td>$3.79</td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  );
+};
+
+export default CSSResetSimpleTablesExample;