+++ linking-platform/link-create/src/index.ts
@@ -1,4 +1,5 @@
-export type { LinkCreatePlugin } from './ui/types';
 export { default, TextField, Select } from './ui/index';
 
 export { useLinkCreateCallback } from './controllers/callback-context/index';
+
+export type { LinkCreateProps, LinkCreatePlugin } from './ui/types';