+++ linking-platform/link-picker/src/index.ts
@@ -2,6 +2,7 @@ export { ComposedLinkPicker as LinkPicker } from './ui';
 export type { LinkPickerProps } from './ui/link-picker';
 export type {
   LinkInputType,
+  LinkPickerPluginAction,
   LinkPickerPlugin,
   LinkPickerState,
   LinkSearchListItemData,